Move MiniMapRenderer to map namespace
